Fix: dividing instantInfo in atoms
